+++ test/LL(1)-parser.cpp
@@ -12,3 +12,14 @@ TEST(LL_1_parser, ShouldAcceptNumericLiteral) {
   //
   // EXPECT_TRUE(accept);
 }
+
+TEST(LL_1_parser, ShouldAcceptStatement) {
+  using namespace pile::Parser;
+
+  const auto grammar = Grammar::parse<"assets/test/ll_one_2.test.glc">();
+  LL1 parser(grammar);
+
+  // const auto accept = parser.parse("1");
+  //
+  // EXPECT_TRUE(accept);
+}