+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.3] - 2022-10-29
 ### Added
-- LL(1) follow set tests
+- LL(1) first set tests
+- LL(1) follow basic set tests
 
 ## [0.1.2] - 2022-10-29
 ### Added